compute/shared/requirements.txt
psycopg2-binary==2.9.9
google-cloud-storage==2.14.2
